solana_message
Module v0

A future Solana message format.

This crate defines two versions of Message in their own modules: legacy and v0. legacy is the current version as of Solana 1.10.0. v0 is a future message format that encodes more account keys into a transaction than the legacy format.

Structs

LoadedAddresses

Collection of addresses loaded from on-chain lookup tables, split by readonly and writable.

LoadedMessage

Combination of a version #0 message and its loaded addresses

Message

A Solana transaction message (v0).

MessageAddressTableLookup

Address table lookups describe an on-chain address lookup table to use for loading more readonly and writable accounts in a single tx.